as listas foram dar uma volta ninguem as quer sao lixo como tu que estas a ler tmb es

+++ POO/src/CarroEletrico.java
@@ -12,7 +12,7 @@ public class CarroEletrico extends Veiculo {
     //-------------------------------------------------------//  
 
     public CarroEletrico (String nova_matricula,
-    					  int velMedia,
+    					  double velMedia,
     					  double priceKM,
     					  double classiF,
     					  Localizacao nova_local,
